Create a "Lootable" component and "Loot Generator"
Create a "Interactive" interface for GameObject that the player could
interact with. From that interface, a "Lootable" has been created that
add items to the player inventory when a GameObject in the scene is
interacted with.
Implement a loot generator from a list of item with weights. Items with
a greater weight have a better chance to get drawn from the loot pool.